working on expand

+++ Parsing/Identificators/identity_tools/identool4.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   identool4.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
+/*   By: ayel-bou <ayel-bou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/04 03:19:33 by ayel-bou          #+#    #+#             */
-/*   Updated: 2025/08/07 06:32:53 by codespace        ###   ########.fr       */
+/*   Updated: 2025/08/14 21:21:48 by ayel-bou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,7 +71,7 @@ static int	open_heredoc(t_token *id_class, t_token *curr, t_data *data)
 	char	*del;
 	char	*gename;
 
-	if (!sef_doc(id_class, data, HERE_SEF))
+	if (!sef_doc(id_class, data))
 		return (0);
 	gename = name_generator();
 	if (!gename)
@@ -82,6 +82,9 @@ static int	open_heredoc(t_token *id_class, t_token *curr, t_data *data)
 	data->here_fd = open(gename, O_CREAT | O_WRONLY, 0777);
 	if (data->here_fd == -1)
 		return (free(del), free(gename), 0);
+	data->here_read_fd = open(gename, O_RDONLY, 0777);
+	if (data->here_read_fd == -1)
+		return (close(data->here_fd), free(del), free(gename), 0);
 	unlink(gename);
 	if (!here_doc_ops(id_class, data, del))
 		return (free(del), free(gename), 0);
